changement py

+++ view/offres.view.php
@@ -14,14 +14,15 @@ include 'include/header.inc.php';
 		<?php 
 		
 		echo "Liste des offres <br/><br/>";
-		
-		foreach($aListeOffres as $k=>$v){
-			echo $v['DESCRIPTIF']; 
-			if ($_SESSION['droit']==1) {
-				?> <a href="index.php?category=21&&id=<?php echo $v['ID_OFFRE'];?>">Modifier Offre</a>
-				<a href="index.php?category=27&&id=<?php echo $v['ID_OFFRE'];?>" onclick="return confirm('Voulez-vous vraiment suprimer cette offre ?');">Supprimer Offre</a>
-				<?php }
-			echo '<br/>';
+		if(!empty($aListeOffres)){
+			foreach($aListeOffres as $k=>$v){
+				echo $v['DESCRIPTIF']; 
+				if ($_SESSION['droit']==1) {
+					?> <a href="index.php?category=21&&id=<?php echo $v['ID_OFFRE'];?>">Modifier Offre</a>
+					<a href="index.php?category=27&&id=<?php echo $v['ID_OFFRE'];?>" onclick="return confirm('Voulez-vous vraiment suprimer cette offre ?');">Supprimer Offre</a>
+					<?php }
+				echo '<br/>';
+			}
 		}
 		
 		if($_SESSION['droit']==1){
